chore: group tests and remove unused packages
