+++ .jest/setup/fetchMock.ts
@@ -1,0 +1,4 @@
+const fetchMock = jest.fn()
+global.fetch = fetchMock
+
+export { fetchMock }
