+++ src/__tests__/index.test.ts
@@ -1,43 +1,7 @@
 import { ApiClient } from '../index'
-
-const fetchMock = jest.fn()
-global.fetch = fetchMock
-
-const mockResolvedOnce = (value: any) => {
-  fetchMock.mockReturnValueOnce(
-    Promise.resolve({
-      ok: true,
-      json: () => {
-        return Promise.resolve(value)
-      },
-      headers: {
-        get: () => {
-          return 200
-        },
-      },
-    }),
-  )
-}
-
-const mockFetchFail = () => {
-  fetchMock.mockReturnValueOnce(Promise.reject(new Error()))
-}
-
-const mockApiFail = () => {
-  fetchMock.mockReturnValueOnce(
-    Promise.resolve({
-      ok: false,
-      statusText: 'Wrong data format',
-      status: 400,
-    }),
-  )
-}
+import { mockFetchFailOnce, mockResolvedOnce } from '../../.jest/helpers/fetch'
 
 describe('ApiClient', () => {
-  beforeEach(() => {
-    fetchMock.mockReset()
-  })
-
   it('throws if there is no baseUrl', () => {
     expect(() => {
       ApiClient.get<{ data: string }>('/listings/search')
@@ -47,7 +11,7 @@ describe('ApiClient', () => {
   })
 
   it('throws if fetch fails', async () => {
-    mockFetchFail()
+    mockFetchFailOnce()
     await expect(
       ApiClient.get<string>('/listings/search', { baseUrl: 'url' }),
     ).rejects.toThrow()
@@ -123,298 +87,5 @@ describe('ApiClient', () => {
         headers: expect.objectContaining({ Authorization: 'Bearer abcdef' }),
       }),
     )
-  })
-
-  describe('get', () => {
-    beforeEach(() => {
-      ApiClient.configure({
-        baseUrl: 'https://api.automotive.ch/api',
-      })
-    })
-
-    it('calls fetch with GET', async () => {
-      mockResolvedOnce({ data: '12345' })
-      await ApiClient.get<{ data: string }>('/listings/search')
-      expect(fetch).toHaveBeenCalledWith(
-        'https://api.automotive.ch/api/listings/search',
-        expect.objectContaining({ method: 'GET' }),
-      )
-    })
-
-    it('extracts the json value of the response', async () => {
-      mockResolvedOnce({ data: '12345' })
-      const data = await ApiClient.get<{ data: string }>('/listings/search')
-      expect(data.data).toEqual('12345')
-    })
-
-    it('rejects if the request was not ok', async () => {
-      mockApiFail()
-      await expect(ApiClient.get<string>('/listings/create')).rejects.toEqual(
-        expect.any(Object),
-      )
-    })
-
-    it('throws if no parameters are passed', () => {
-      expect(() => {
-        ApiClient.get<{ data: string }>(
-          'dealers/{dealerId}/listings/{listingId}',
-          {},
-        )
-      }).toThrow()
-    })
-
-    it('throws if a parameter is missing', async () => {
-      expect(() => {
-        ApiClient.get<{ data: string }>(
-          'dealers/{dealerId}/listings/{listingId}',
-          {
-            params: { listingId: 456 },
-          },
-        )
-      }).toThrow(
-        'Param {dealerId} missing. Expected parameters are: {dealerId}, {listingId}',
-      )
-    })
-
-    it('replaces all occurrences if the parameters are named the same', async () => {
-      mockResolvedOnce({ data: '12345' })
-      await ApiClient.get<string>('dealers/{dealerId}/listings/{dealerId}', {
-        params: { dealerId: 123 },
-      })
-      expect(fetch).toHaveBeenCalledWith(
-        expect.stringContaining('dealers/123/listings/123'),
-        expect.any(Object),
-      )
-    })
-
-    it('replaces the parameter placeholders with the data', async () => {
-      mockResolvedOnce({ data: '12345' })
-      await ApiClient.get<string>('dealers/{dealerId}/listings/{listingId}', {
-        params: { dealerId: 123, listingId: 456 },
-      })
-      expect(fetch).toHaveBeenCalledWith(
-        expect.stringContaining('dealers/123/listings/456'),
-        expect.any(Object),
-      )
-    })
-  })
-
-  describe('post', () => {
-    beforeEach(() => {
-      ApiClient.configure({
-        baseUrl: 'https://api.automotive.ch/api',
-      })
-    })
-
-    it('calls fetch with POST', async () => {
-      mockResolvedOnce({ id: '12345' })
-      await ApiClient.post<{ id: string }, { make: string }>(
-        '/listings/create',
-        { make: 'bmw' },
-      )
-      expect(fetch).toHaveBeenCalledWith(
-        'https://api.automotive.ch/api/listings/create',
-        expect.objectContaining({ method: 'POST' }),
-      )
-    })
-
-    it('extracts the json value of the response', async () => {
-      mockResolvedOnce({ id: '12345' })
-      const data = await ApiClient.post<{ id: string }, { make: string }>(
-        '/listings/create',
-        { make: 'bmw' },
-      )
-      expect(data.id).toEqual('12345')
-    })
-
-    it('rejects if the request was not ok', async () => {
-      mockApiFail()
-      await expect(
-        ApiClient.post<string, string>('/listings/create', 'data'),
-      ).rejects.toEqual(expect.any(Object))
-    })
-
-    it('throws if no parameters are passed', () => {
-      expect(() => {
-        ApiClient.post<{ id: string }, { make: string }>(
-          'dealers/{dealerId}/listings/{listingId}',
-          { make: 'bmw' },
-          {},
-        )
-      }).toThrow()
-    })
-
-    it('throws if a parameter is missing', async () => {
-      expect(() => {
-        ApiClient.post<{ id: string }, { make: string }>(
-          'dealers/{dealerId}/listings/{listingId}',
-          { make: 'bmw' },
-          {
-            params: { listingId: 456 },
-          },
-        )
-      }).toThrow(
-        'Param {dealerId} missing. Expected parameters are: {dealerId}, {listingId}',
-      )
-    })
-
-    it('replaces the parameter placeholders with the data', async () => {
-      mockResolvedOnce({ data: '12345' })
-      await ApiClient.post<{ id: string }, { make: string }>(
-        'dealers/{dealerId}/listings/{listingId}',
-        { make: 'bmw' },
-        {
-          params: { dealerId: 123, listingId: 456 },
-        },
-      )
-      expect(fetch).toHaveBeenCalledWith(
-        expect.stringContaining('dealers/123/listings/456'),
-        expect.any(Object),
-      )
-    })
-
-    it('calls fetch with an empty body', async () => {
-      mockResolvedOnce(null)
-      await ApiClient.post<{ id: string }, null>(
-        'dealers/{dealerId}/listings/{listingId}',
-        null,
-        {
-          params: { dealerId: 123, listingId: 456 },
-        },
-      )
-      expect(fetch).toHaveBeenCalledWith(
-        expect.stringContaining('dealers/123/listings/456'),
-        expect.objectContaining({ body: null }),
-      )
-    })
-  })
-
-  describe('put', () => {
-    beforeEach(() => {
-      ApiClient.configure({
-        baseUrl: 'https://api.automotive.ch/api',
-      })
-    })
-
-    it('calls fetch with PUT', async () => {
-      mockResolvedOnce({ id: '12345' })
-      await ApiClient.put<string, { make: string }>('/listings/update', {
-        make: 'bmw',
-      })
-      expect(fetch).toHaveBeenCalledWith(
-        'https://api.automotive.ch/api/listings/update',
-        expect.objectContaining({ method: 'PUT' }),
-      )
-    })
-
-    it('extracts the json value of the response', async () => {
-      mockResolvedOnce({ id: '12345' })
-      const data = await ApiClient.put<{ id: string }, { make: string }>(
-        '/listings/update',
-        { make: 'bmw' },
-      )
-      expect(data.id).toEqual('12345')
-    })
-
-    it('rejects if the request was not ok', async () => {
-      mockApiFail()
-      await expect(
-        ApiClient.put<string, string>('/listings/update', 'data'),
-      ).rejects.toEqual(expect.any(Object))
-    })
-
-    it('throws if no parameters are passed', () => {
-      expect(() => {
-        ApiClient.put<{ id: string }, { make: string }>(
-          'dealers/{dealerId}/listings/{listingId}',
-          { make: 'bmw' },
-          {},
-        )
-      }).toThrow()
-    })
-
-    it('throws if a parameter is missing', async () => {
-      expect(() => {
-        ApiClient.put<{ id: string }, { make: string }>(
-          'dealers/{dealerId}/listings/{listingId}',
-          { make: 'bmw' },
-          {
-            params: { listingId: 456 },
-          },
-        )
-      }).toThrow(
-        'Param {dealerId} missing. Expected parameters are: {dealerId}, {listingId}',
-      )
-    })
-
-    it('replaces the parameter placeholders with the data', async () => {
-      mockResolvedOnce({ data: '12345' })
-      await ApiClient.put<{ id: string }, { make: string }>(
-        'dealers/{dealerId}/listings/{listingId}',
-        { make: 'bmw' },
-        {
-          params: { dealerId: 123, listingId: 456 },
-        },
-      )
-      expect(fetch).toHaveBeenCalledWith(
-        expect.stringContaining('dealers/123/listings/456'),
-        expect.any(Object),
-      )
-    })
-  })
-
-  describe('delete', () => {
-    beforeEach(() => {
-      ApiClient.configure({
-        baseUrl: 'https://api.automotive.ch/api',
-      })
-    })
-
-    it('calls fetch with DELETE', async () => {
-      mockResolvedOnce({ data: '12345' })
-      await ApiClient.delete('/listings/{listingId}', {
-        params: { listingId: 123 },
-      })
-      expect(fetch).toHaveBeenCalledWith(
-        expect.any(String),
-        expect.objectContaining({ method: 'DELETE' }),
-      )
-    })
-
-    it('rejects if the request was not ok', async () => {
-      mockApiFail()
-      await expect(
-        ApiClient.delete('/listings/{listingId}', {
-          params: { listingId: 123 },
-        }),
-      ).rejects.toEqual(expect.any(Object))
-    })
-
-    it('throws if no parameters are passed', () => {
-      expect(() => {
-        ApiClient.delete('dealers/{dealerId}/listings/{listingId}', {})
-      }).toThrow()
-    })
-
-    it('throws if a parameter is missing', async () => {
-      expect(() => {
-        ApiClient.delete('dealers/{dealerId}/listings/{listingId}', {
-          params: { listingId: 456 },
-        })
-      }).toThrow(
-        'Param {dealerId} missing. Expected parameters are: {dealerId}, {listingId}',
-      )
-    })
-
-    it('replaces the parameter placeholders with the data', async () => {
-      mockResolvedOnce({})
-      await ApiClient.delete('dealers/{dealerId}/listings/{listingId}', {
-        params: { dealerId: 123, listingId: 456 },
-      })
-      expect(fetch).toHaveBeenCalledWith(
-        expect.stringContaining('dealers/123/listings/456'),
-        expect.any(Object),
-      )
-    })
   })
 })
